Modify to throw a message when an error occurs in asynchronous processing
