--- src/assets/mixin/main-btn.scss
@@ -1,21 +1,0 @@
-@import "../variable";
-
-@mixin main-btn {
-  background-color: white;
-  border: 2px solid $color-geyser ;
-  border-radius: $task-border_radius;
-  color: $color-lynch;
-  cursor: pointer;
-  font-size: 1.0rem;
-  padding: 5px 20px 5px 20px;
-
-  &:hover{
-    background-color: $color-lynch;
-    border: 2px solid $color-geyser ;
-    color: white;
-  }
-
-  &:focus {
-    outline: none;
-  }
-}
