+++ src/components/modules/inputYears/InputYears.module.scss
@@ -10,10 +10,16 @@
 }
 
 .prevBtn {
-  @extend .task--btn;
+  border: 1.5px solid $color-geyser ;
   border-radius: 0.5rem 0 0 0.5rem;
   border-right: none;
-  padding-top: 8px;
+  color: $color-scorpion;
+  cursor: pointer;
+  padding: 8px 13px 3px;
+
+  &:focus {
+    outline: none;
+  }
 }
 
 .jumpYears {
@@ -38,14 +44,19 @@
 }
 
 .nextBtn {
-  @extend .task--btn;
+  border: 1.5px solid $color-geyser ;
   border-radius: 0 0.5rem 0.5rem 0;
   border-left: none;
-  padding-top: 8px;
+  color: $color-scorpion;
+  cursor: pointer;
+  padding: 8px 13px 3px;
+
+  &:focus {
+    outline: none;
+  }
 }
 
 .wrapper {
-  background-color: $color-porcelain;
   border-radius: 0.5rem;
   border: $color-geyser solid 1px;
   color: $color-tundora;
