+++ src/components/todo/modules/area/searchTodoListArea/SearchTodoListArea.module.scss
@@ -1,0 +1,84 @@
+@import "../../../../../assets/mixin/icon-btn";
+@import "../../../../../assets/mixin/operation-btn";
+
+.wrapper {
+  width: 70%;
+  background-color: white;
+  border: $color-geyser solid 1px;
+  border-radius: 0.5rem;
+  padding: 0 0 20px;
+  margin: 0 auto;
+  box-shadow: $section-box_shadow;
+}
+
+.position {
+  display: flex;
+  border-top-right-radius: 0.5rem;
+  border-top-left-radius: 0.5rem;
+  border-bottom: $color-geyser solid 1px;
+  justify-content: center;
+  position: relative;
+  background-color: $color-blackSqueeze;
+  padding: 10px 0 10px 0;
+  margin: 0 0 25px 0;
+}
+
+.btnPosition{
+  @include icon-btn($color-scorpion, 0.9rem, 0.5);
+  font-size: 1.5rem;
+  position: absolute;
+  left: 0;
+  margin: 0 0 0 20px;
+}
+
+.title {
+  color: $color-scorpion;
+  font-size: 1.0rem;
+  margin: 0;
+}
+
+.selectItems {
+  display: flex;
+  margin: 0 0 30px 0;
+
+  &__key {
+    width: 180px;
+    color: $color-scorpion;
+    font-size: 1.0rem;
+    font-weight: 500;
+    margin: 0 0 0 20px;
+  }
+
+  &__value {
+    color: $color-scorpion;
+  }
+}
+
+.dateDesignation{
+  display: flex;
+  align-items: center;
+}
+
+.datePickers {
+  display: flex;
+  align-items: center;
+  margin-left: 50px;
+
+  &__space {
+    margin: 0 20px;
+  }
+}
+
+.searchBtnPosition {
+  text-align: center;
+  margin-bottom: 20px
+}
+
+.searchBtn {
+  @include operation-btn($color-dodgerBlue, white, $color-dodgerBlue);
+
+  &:hover {
+    background-color: $color-denim;
+    border-color: $color-denim;
+  }
+}
