+++ src/components/todo/modules/area/searchTodoListArea/searchResultTodoList/SarchResultTodoList.module.scss
@@ -1,0 +1,45 @@
+@import "../../../../../../assets/variable";
+
+.wrapper {
+  border-radius: 0.6rem;
+  margin: 0 20px;
+  box-shadow: $list-item-box_shadow;
+}
+
+.listByDate {
+  width: 60%;
+  list-style-type: none;
+  padding: 20px 0;
+  margin: 0 auto;
+
+  > div, li {
+    margin-bottom: 20px;
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
+
+.date {
+  color: $color-scorpion;
+  font-size: 1.0rem;
+  margin: 30px 0 20px 0;
+
+  &:first-child {
+    margin: 0 0 20px 0;
+  }
+}
+
+.message {
+  color: $color-scorpion;
+  font-size: 0.95rem;
+  font-weight: 500;
+  text-align: center;
+  padding: 25px;
+  margin: 0;
+}
+
+.childFormClassName {
+  width: 70%;
+}
