+++ src/components/todo/modules/area/todayTodoListArea/TodayTodoListArea.module.scss
@@ -1,0 +1,19 @@
+@import "../../../../../assets/variable";
+
+.addButton {
+  margin: 20px 0;
+}
+
+.switchItem {
+  padding: 25px 0;
+  height: auto;
+
+  &__width {
+    width: 60%;
+    margin: auto;
+  }
+}
+
+.childTabsBtn {
+  margin-bottom: 30px;
+}
