--- src/components/todo/page/ExpiredTodoListArea/ExpiredTodoListArea.module.scss
@@ -1,76 +1,0 @@
-@import "../../../../assets/variable";
-
-.listByDate {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.listItemByDate {
-  margin-bottom: 30px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-}
-
-.date {
-  color: $color-scorpion;
-  font-size: 1.0rem;
-  margin: 0 0 20px 0;
-}
-
-.todoList {
-  list-style-type: none;
-  padding: 0;
-
-  > div, li {
-    margin-bottom: 20px;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-}
-
-.readMoreBtn {
-  display: flex;
-  align-items: center;
-  background-color: $color-denim;
-  color: white;
-  font-size: 0.8rem;
-  border: none;
-  border-radius: 0.6rem;
-  padding: 2px 5px 2px 15px;
-  cursor: pointer;
-  outline: none;
-
-  &:hover {
-    opacity: 0.9;
-  }
-}
-
-.message {
-  color: $color-scorpion;
-  font-size: 0.95rem;
-  font-weight: 500;
-  text-align: center;
-  padding: 10px 0;
-  margin: 0;
-}
-
-.childListItem {
-  margin-bottom: 20px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-}
-
-.childListItemForm {
-  margin-bottom: 20px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-}
