--- src/components/todo/page/MonthlyTodoListArea/MonthlyTodoListArea.module.scss
@@ -1,32 +1,0 @@
-@import "../../../../assets/variable";
-
-.addButton {
-  margin: 20px 0;
-}
-
-.switchItem {
-  border-radius: 0.6rem;
-  padding: 25px 0;
-  height: auto;
-
-  &__width {
-    width: 60%;
-    margin: auto;
-  }
-}
-
-.inputYears {
-  margin-bottom: 20px;
-}
-
-.childInputYearsBtn {
-  background-color: $color-porcelain;
-
-  &:hover{
-    background-color: $color-athensGray;
-  }
-}
-
-.childTabsBtn {
-  margin-bottom: 30px;
-}
