--- src/components/todo/page/TodayTodoListArea/TodayTodoListArea.tsx
@@ -1,51 +1,0 @@
-import React from 'react';
-import SwitchItemTabs from '../../../uikit/tabs/switchItemTabs/SwitchItemTabs';
-import TodayImplementationDateTodoListContainer from '../../../../containers/todo/page/todayTodoListArea/Items/TodayImplementationDateTodoListContainer';
-import TodayDueDateTodoListContainer from '../../../../containers/todo/page/todayTodoListArea/Items/TodayDueDateTodoListContainer';
-import styles from './TodayTodoListArea.module.scss';
-import AddTodoListItemFormContainer from '../../../../containers/todo/modules/form/AddTodoListItemFormContainer';
-
-interface TodayTodoListAreaProps {
-  currentYear: string;
-  currentMonth: string;
-  editing: boolean;
-  setEditing: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-const TodayTodoListArea = (props: TodayTodoListAreaProps) => {
-  return (
-    <>
-      <div className={styles.addButton}>
-        <AddTodoListItemFormContainer
-          currentYear={props.currentYear}
-          currentMonth={props.currentMonth}
-        />
-      </div>
-      <div className={styles.switchItem}>
-        <div className={styles.switchItem__width}>
-          <SwitchItemTabs
-            leftButtonLabel={'実施予定のToDo'}
-            rightButtonLabel={'締切予定のToDo'}
-            leftItem={
-              <TodayImplementationDateTodoListContainer
-                currentYear={props.currentYear}
-                currentMonth={props.currentMonth}
-                setEditing={props.setEditing}
-              />
-            }
-            rightItem={
-              <TodayDueDateTodoListContainer
-                currentYear={props.currentYear}
-                currentMonth={props.currentMonth}
-                setEditing={props.setEditing}
-              />
-            }
-            tabsBtnClassName={styles.childTabsBtn}
-          />
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default TodayTodoListArea;
