+++ src/reducks/groupTodoList/operations.ts
@@ -235,7 +235,7 @@ export const addGroupTodoListItem = (
         }
       );
 
-      const fetchExpiredResult = await axios.get<FetchGroupExpiredTodoListRes>(
+      const fetchExpiredResult = axios.get<FetchGroupExpiredTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/expired`,
         {
           cancelToken: signal.token,
@@ -243,7 +243,7 @@ export const addGroupTodoListItem = (
         }
       );
 
-      const fetchTodayResult = await axios.get<FetchGroupTodayTodoListRes>(
+      const fetchTodayResult = axios.get<FetchGroupTodayTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${year}-${month}-${date}`,
         {
           cancelToken: signal.token,
@@ -251,7 +251,7 @@ export const addGroupTodoListItem = (
         }
       );
 
-      const fetchMonthlyResult = await axios.get<FetchGroupMonthlyTodoListRes>(
+      const fetchMonthlyResult = axios.get<FetchGroupMonthlyTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${currentYear}-${currentMonth}`,
         {
           cancelToken: signal.token,
@@ -259,18 +259,23 @@ export const addGroupTodoListItem = (
         }
       );
 
-      const todayMessage = fetchTodayResult.data.message;
-      const monthlyMessage = fetchMonthlyResult.data.message;
+      const expiredResponse = await fetchExpiredResult;
+      const todayResponse = await fetchTodayResult;
+      const monthlyResponse = await fetchMonthlyResult;
 
-      const expiredTodoList = fetchExpiredResult.data.expired_group_todo_list;
+      const expiredTodoList = expiredResponse.data.expired_group_todo_list;
+
+      const todayMessage = todayResponse.data.message;
       const todayImplementationTodoList = todayMessage
         ? []
-        : fetchTodayResult.data.implementation_todo_list;
-      const todayDuesTodoList = todayMessage ? [] : fetchTodayResult.data.due_todo_list;
+        : todayResponse.data.implementation_todo_list;
+      const todayDuesTodoList = todayMessage ? [] : todayResponse.data.due_todo_list;
+
+      const monthlyMessage = monthlyResponse.data.message;
       const monthlyImplementationTodoList = monthlyMessage
         ? []
-        : fetchMonthlyResult.data.implementation_todo_list;
-      const monthlyDuesTodoList = monthlyMessage ? [] : fetchMonthlyResult.data.due_todo_list;
+        : monthlyResponse.data.implementation_todo_list;
+      const monthlyDuesTodoList = monthlyMessage ? [] : monthlyResponse.data.due_todo_list;
 
       dispatch(
         addGroupTodoListItemAction(
@@ -285,6 +290,7 @@ export const addGroupTodoListItem = (
       dispatch(
         failedAddGroupTodoListItemAction(error.response.status, error.response.data.error.message)
       );
+      throw error.response.data.error.message;
     }
   };
 };
@@ -319,7 +325,7 @@ export const editGroupTodoListItem = (
         }
       );
 
-      const fetchExpiredResult = await axios.get<FetchGroupExpiredTodoListRes>(
+      const fetchExpiredResult = axios.get<FetchGroupExpiredTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/expired`,
         {
           cancelToken: signal.token,
@@ -327,7 +333,7 @@ export const editGroupTodoListItem = (
         }
       );
 
-      const fetchTodayResult = await axios.get<FetchGroupTodayTodoListRes>(
+      const fetchTodayResult = axios.get<FetchGroupTodayTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${year}-${month}-${date}`,
         {
           cancelToken: signal.token,
@@ -335,7 +341,7 @@ export const editGroupTodoListItem = (
         }
       );
 
-      const fetchMonthlyResult = await axios.get<FetchGroupMonthlyTodoListRes>(
+      const fetchMonthlyResult = axios.get<FetchGroupMonthlyTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${currentYear}-${currentMonth}`,
         {
           cancelToken: signal.token,
@@ -343,18 +349,23 @@ export const editGroupTodoListItem = (
         }
       );
 
-      const todayMessage = fetchTodayResult.data.message;
-      const monthlyMessage = fetchMonthlyResult.data.message;
+      const expiredResponse = await fetchExpiredResult;
+      const todayResponse = await fetchTodayResult;
+      const monthlyResponse = await fetchMonthlyResult;
 
-      const expiredTodoList = fetchExpiredResult.data.expired_group_todo_list;
+      const expiredTodoList = expiredResponse.data.expired_group_todo_list;
+
+      const todayMessage = todayResponse.data.message;
       const todayImplementationTodoList = todayMessage
         ? []
-        : fetchTodayResult.data.implementation_todo_list;
-      const todayDuesTodoList = todayMessage ? [] : fetchTodayResult.data.due_todo_list;
+        : todayResponse.data.implementation_todo_list;
+      const todayDuesTodoList = todayMessage ? [] : todayResponse.data.due_todo_list;
+
+      const monthlyMessage = monthlyResponse.data.message;
       const monthlyImplementationTodoList = monthlyMessage
         ? []
-        : fetchMonthlyResult.data.implementation_todo_list;
-      const monthlyDuesTodoList = monthlyMessage ? [] : fetchMonthlyResult.data.due_todo_list;
+        : monthlyResponse.data.implementation_todo_list;
+      const monthlyDuesTodoList = monthlyMessage ? [] : monthlyResponse.data.due_todo_list;
 
       dispatch(
         editGroupTodoListItemAction(
@@ -369,6 +380,7 @@ export const editGroupTodoListItem = (
       dispatch(
         failedEditGroupTodoListItemAction(error.response.status, error.response.data.error.message)
       );
+      throw error.response.data.error.message;
     }
   };
 };
@@ -394,7 +406,7 @@ export const deleteGroupTodoListItem = (
         }
       );
 
-      const fetchExpiredResult = await axios.get<FetchGroupExpiredTodoListRes>(
+      const fetchExpiredResult = axios.get<FetchGroupExpiredTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/expired`,
         {
           cancelToken: signal.token,
@@ -402,7 +414,7 @@ export const deleteGroupTodoListItem = (
         }
       );
 
-      const fetchTodayResult = await axios.get<FetchGroupTodayTodoListRes>(
+      const fetchTodayResult = axios.get<FetchGroupTodayTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${year}-${month}-${date}`,
         {
           cancelToken: signal.token,
@@ -410,7 +422,7 @@ export const deleteGroupTodoListItem = (
         }
       );
 
-      const fetchMonthlyResult = await axios.get<FetchGroupMonthlyTodoListRes>(
+      const fetchMonthlyResult = axios.get<FetchGroupMonthlyTodoListRes>(
         `${process.env.REACT_APP_TODO_API_HOST}/groups/${groupId}/todo-list/${currentYear}-${currentMonth}`,
         {
           cancelToken: signal.token,
@@ -418,18 +430,23 @@ export const deleteGroupTodoListItem = (
         }
       );
 
-      const todayMessage = fetchTodayResult.data.message;
-      const monthlyMessage = fetchMonthlyResult.data.message;
+      const expiredResponse = await fetchExpiredResult;
+      const todayResponse = await fetchTodayResult;
+      const monthlyResponse = await fetchMonthlyResult;
 
-      const expiredTodoList = fetchExpiredResult.data.expired_group_todo_list;
+      const expiredTodoList = expiredResponse.data.expired_group_todo_list;
+
+      const todayMessage = todayResponse.data.message;
       const todayImplementationTodoList = todayMessage
         ? []
-        : fetchTodayResult.data.implementation_todo_list;
-      const todayDuesTodoList = todayMessage ? [] : fetchTodayResult.data.due_todo_list;
+        : todayResponse.data.implementation_todo_list;
+      const todayDuesTodoList = todayMessage ? [] : todayResponse.data.due_todo_list;
+
+      const monthlyMessage = monthlyResponse.data.message;
       const monthlyImplementationTodoList = monthlyMessage
         ? []
-        : fetchMonthlyResult.data.implementation_todo_list;
-      const monthlyDuesTodoList = monthlyMessage ? [] : fetchMonthlyResult.data.due_todo_list;
+        : monthlyResponse.data.implementation_todo_list;
+      const monthlyDuesTodoList = monthlyMessage ? [] : monthlyResponse.data.due_todo_list;
 
       dispatch(
         deleteGroupTodoListItemAction(
@@ -448,6 +465,7 @@ export const deleteGroupTodoListItem = (
           error.response.data.error.message
         )
       );
+      throw error.response.data.error.message;
     }
   };
 };
@@ -506,6 +524,7 @@ export const editGroupSearchTodoListItem = (
           error.response.data.error.message
         )
       );
+      throw error.response.data.error.message;
     }
   };
 };
@@ -556,6 +575,7 @@ export const deleteGroupSearchTodoListItem = (
           error.response.data.error.message
         )
       );
+      throw error.response.data.error.message;
     }
   };
 };
